+++ rmf_site_editor/src/widgets/icons.rs
@@ -32,13 +32,13 @@ pub struct Icons {
 impl FromWorld for Icons {
     fn from_world(world: &mut World) -> Self {
         let asset_server = world.get_resource::<AssetServer>().unwrap();
-        let bevy_select = asset_server.load(&String::from(AssetSource::Remote(
+        let bevy_select = asset_server.load(&String::from(&AssetSource::Remote(
             "textures/select.png".to_string(),
         )));
-        let bevy_edit = asset_server.load(&String::from(AssetSource::Remote(
+        let bevy_edit = asset_server.load(&String::from(&AssetSource::Remote(
             "textures/edit.png".to_string(),
         )));
-        let bevy_trash = asset_server.load(&String::from(AssetSource::Remote(
+        let bevy_trash = asset_server.load(&String::from(&AssetSource::Remote(
             "textures/trash.png".to_string(),
         )));
